pyproject.toml
[project]
name = "movies-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel",
    "alembic",
    "pandas",
    "python-dotenv",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain[google-genai]>=0.3.25",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.3",
    "pytest>=8.4.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


